+++ docs/guides/changelog.md
@@ -5,6 +5,7 @@ page_title: "Changelog"
 # Changelog
 
 ## v1.5.0
+
 * Add support for managing the AWS Outpost account as a separate `polaris_aws_account` resource. The
   `outpost_account_id` and `outpost_account_profile` fields have been made optional.
 * Add support for the following feature blocks in the `polaris_aws_account` resource: `cloud_discovery`,
@@ -17,6 +18,16 @@ page_title: "Changelog"
 * Remove the in-place update functionality from the `polaris_aws_exocompute` resource due to API issues. The in-place
   update functionality was only used for the `cluster_access` field. Updating the field now requires the resource to be
   re-created.
+* Add support for user-assigned managed identity in the `polaris_azure_subscription` resource for the SQL DB Protection
+  feature. The managed identity fields (`user_assigned_managed_identity_name`, `user_assigned_managed_identity_principal_id`,
+  `user_assigned_managed_identity_region`, and `user_assigned_managed_identity_resource_group_name`) can be specified
+  directly in the `sql_db_protection` block. This is required when using Transparent Data Encryption (TDE) with customer
+  managed keys. [[docs](../resources/azure_subscription.md#nested-schema-for-sql_db_protection)]
+  directly in the `sql_db_protection` block. These fields are required once the RSC account has the
+  `CNP_AZURE_SQL_DB_TDE_CMK_SUPPORT` feature flag enabled for Transparent Data Encryption (TDE) with customer managed keys.
+  Specifying these fields before the feature flag is enabled will result in an error. Supports upgrade scenarios where the
+  feature flag is enabled on existing SQL DB Protection configurations.
+  [[docs](../resources/azure_subscription.md#nested-schema-for-sql_db_protection)]
 
 ## v1.4.0